Add project
In this dataset, there are some wrong values ​​in the Area column and their data type is string. I replaced the wrong values ​​with nan and changed the data type. I labeled some of the columns with one hot encoder to find their correlation, which I found to be not very well correlated.Then I examined the columns that had good correlations separately and saw that it was not possible to obtain a very accurate prediction from them. So, using Polynomial, I combined two of the columns into one equation and the prediction became better.